Remove LuaBinding and refactor engine bindings
LuaBinding has been deleted and its responsibilities moved into EngineBinding, which now directly manages Lua audio bindings. AppView and EngineBinding interfaces have been updated to remove LuaBinding references and unify key code definitions. Various Vulkan and engine initialization routines have been refactored for clarity and correctness, and minor fixes were applied to gfx and pipeline logic. Project configuration files updated for new development environment and build settings.

+++ src/tknCore.c
@@ -157,7 +157,7 @@ void *tknGetFromDynamicArray(TknDynamicArray *pDynamicArray, uint32_t index)
     }
 }
 
-TknHashSet tknCreateHashSet(size_t capacity)
+TknHashSet tknCreateHashSet(uint32_t capacity)
 {
     TknHashSet tknHashSet = {
         .capacity = capacity,
@@ -183,7 +183,7 @@ bool tknAddToHashSet(TknHashSet *pTknHashSet, void *value)
         if (pTknHashSet->count * 10 >= pTknHashSet->capacity * 8)
         {
             // Resize the hash set
-            size_t newCapacity = pTknHashSet->capacity * 2;
+            uint32_t newCapacity = pTknHashSet->capacity * 2;
             TknListNode **newNodePtrs = tknMalloc(sizeof(TknListNode *) * newCapacity);
             memset(newNodePtrs, 0, sizeof(TknListNode *) * newCapacity);
             for (size_t i = 0; i < pTknHashSet->capacity; i++)
